test: move ECH GREASE regression test to test/regression/issue/29780.test.ts
Avoids coupling to a flaky timing-sensitive neighbor in fetch.tls.test.ts
('fetch timeout works on tls') that fails intermittently on main.
